Use TLongObjectHashMap instead of THashMap. (#37)
* Use TLongObjectHashMap instead of THashMap.

... in order to save memory. This locks us on using longs which i think
is fine.

* Removed unused imports.

+++ src/main/java/org/apache/tinkerpop/gremlin/tinkergraph/structure/TinkerVertex.java
@@ -138,7 +138,7 @@ public class TinkerVertex extends TinkerElement implements Vertex {
         }).forEach(Edge::remove);
         this.properties = null;
         TinkerHelper.removeElementIndex(this);
-        graph.vertices.remove(id());
+        graph.vertices.remove((long)id());
         graph.getElementsByLabel(graph.verticesByLabel, label).remove(this);
 
         if (graph.ondiskOverflowEnabled) {
